requirements.txt
aiohttp
python-telegram-bot[rate-limiter]
python-dotenv